Extracted some functions into separate files
