+++ include/modloader/utils.h
@@ -26,3 +26,17 @@ extern __declspec(dllexport) void redirectCall(const void* target, const uint64_
 
 extern __declspec(dllexport) const HexToStringResult
     parseByteArray(const std::string& str, std::vector<uint8_t>& vector);
+
+template<typename Input, typename Output, typename Extractor>
+void copy_max(Input& container, Output output, std::size_t max, Extractor func)
+{
+    uint32_t count = 0;
+    for (auto it = container.begin(); count < max && it != container.end(); count++, output++, it++)
+        *output = func(it);
+}
+
+template<typename Input, typename Output, typename Extractor>
+void copy_max(Input* container, Output output, std::size_t max, Extractor func)
+{
+    copy_max(*container, output, max, func);
+}
